revert of test_all.py, update of debugger test
